+++ src/action_flee_end.c
@@ -7,7 +7,7 @@
 
 extern void unit_stop_running_blindly(int unit_index);
 
-void action_flee_end(uint16_t actor_index)
+void action_flee_end(int actor_index)
 {
     actor_datum *actor = DATA_ARRAY_ELEMENT(actor_data, actor_datum, actor_index);
     if (actor->meta.unit_index != -1)
