+++ src/action_flee_flush_position_indices.c
@@ -9,7 +9,7 @@
 #include "headers/blam_data_globals.h"
 
 
-void action_flee_flush_position_indices(uint16_t actor_index)
+void action_flee_flush_position_indices(int actor_index)
 {
     actor_datum *actor = DATA_ARRAY_ELEMENT(actor_data, actor_datum, actor_index);
     actor->state.action_data.___u0.flee.flee_firing_position_index = -1;
